Refactor Relationship and Label into Properties
The bulk of the macro was exactly the same, and the purpose of modules using
the macro is simply to validate the properties, which is applicable to both
Edge and Vertex structures (except the latter can support multiples).  Reflect
that by refactoring into simply "Properties", and then the Vertex can validate
by iterating over the labels (validators), Edge simply does it once for
its type.  Added a call "name" that can be made in properties to avoid it
using a CamelCase'd version of the module's name.  Finished changing
Vertex.cast to Vertex.new.