properties macro accepts id: keyword parameter
By default, an :id field will be added to the struct, and we will note
that it's defined to be a :string; but you can pass an alternative value
on the properties parameters (like parameters id: :integer, do).  If the
alternative value is non-true (nil or false), the id field won't be added
to the struct.  (see properties_test.exs)