Refactor Properties methods to be public.
validate() and validate!() can be used directly, useful if you already
have a Vertex or Edge and want to re-validate the content.  apply_defaults()
applies, well, defaults.  atomize_known_keys does what it says.  This allows
the whole process to be defined differently, or to be applied discrete maps,
or to apply different validators/defaults against a map that isn't yet a
Vertex.  These methods are also defined on the Properties structs so they
can be called with an implicit module as the type.