Adding support for automagically generated IDs.
By default, a field :id will be automagically added to any Properties
struct, and also by default, it will be filled in with a UUID when you
call "new".  If this field exists in a request to get(), all other
properties are ignored, and only the ID is used to find the object.