Change bitstring to binary (n00b confusion)
Not sure why, but I was certain bitstrings were used for, well, strings,
and binary was for unaligned binary data.  But I was wrong.  :P

Also, get_or_create can take a set of properties to be used to create a
record if not otherwise matched -- so you can search on select keys, but
provide a full record if nothing was found.